回复功能，异常处理，将id重构为Long类型

+++ src/main/java/cgh/community/community/dto/QuestionDTO.java
@@ -8,13 +8,13 @@ import lombok.Data;
  */
 @Data
 public class QuestionDTO {
-    private Integer id; //问题id
+    private Long id; //问题id
     private String title;   //标题
     private String description; //描述
     private String tag; //标签
     private Long gmtCreate; //创建时间
     private Long gmtModified;   //修改时间
-    private Integer creator;    //创建问题用户id
+    private Long creator;    //创建问题用户id
     private Integer viewCount;  //浏览数
     private Integer commentCount;   //评论数
     private Integer likeCount;  //点赞数
